Add a broken test

+++ src/Basket.API/Grpc/BasketService.cs
@@ -1,7 +1,8 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using eShop.Basket.API.Repositories;
+﻿using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using eShop.Basket.API.Extensions;
 using eShop.Basket.API.Model;
+using eShop.Basket.API.Repositories;
 
 namespace eShop.Basket.API.Grpc;
 
@@ -62,6 +63,11 @@ public class BasketService(
         if (string.IsNullOrEmpty(userId))
         {
             ThrowNotAuthenticated();
+        }
+        var basket = await repository.GetBasketAsync(userId);
+        if (basket.Items.Any())
+        {
+            throw new InvalidOperationException("Cannot delete a that contains items");
         }
 
         await repository.DeleteBasketAsync(userId);
